PWA-1282: Improved ViewComponent structure
