+++ themes/theme-gmd/components/View/components/Content/style.js
